+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+* `RoundedBackgroundTextField` now scales down to fit the text
+* `RoundedBackgroundTextField` now can't be scrollable
+
 ## 0.1.1
 
 * Customizable radius